+++ packages/repl/src/index.ts
@@ -1,19 +1,21 @@
 import { BaseREPL, BaseREPLContext, CommandREPL, CommandREPLContext } from './repl';
-import TidalREPL from './repl/tidal';
-import SclangREPL, { RemoteSclangREPL } from './repl/sclang';
 import FoxDotREPL from './repl/foxdot';
 import MercuryREPL from './repl/mercury';
+import SclangREPL, { RemoteSclangREPL } from './repl/sclang';
+import SonicPiREPL from './repl/sonicpi';
+import TidalREPL from './repl/tidal';
 
 const path = require("path");
 const fs = require("fs");
 
 const replClasses = {
   default: CommandREPL,
-  tidal: TidalREPL,
-  sclang: SclangREPL,
-  remote_sclang: RemoteSclangREPL,
   foxdot: FoxDotREPL,
   mercury: MercuryREPL,
+  remote_sclang: RemoteSclangREPL,
+  sclang: SclangREPL,
+  sonicpi: SonicPiREPL,
+  tidal: TidalREPL,
 };
 
 function createREPLFor(repl: string, ctx: CommandREPLContext) {